Refactoring done on File Manager.
It shall now be possible to access the file system directly, without
the use of a caching tree or browser. The interface to the file system
has been purified. All functions can now be called using a const char, or
with a combination of a path structure and a const char, whichever is more
convenient.  If special filesystems exist, they can override the function
'walk_path', such that the file open call gets the correct parameters.


git-svn-id: https://xp-dev.com/svn/1541UltimateII@248 6fa36889-7ba6-4ba5-89ae-355e570fd296