FATFS re-enabled reentrancy protection, and tested for unbalanced request/release mutex calls by using a dummy. Seems OK with this limited test case.
Note: Protection is not yet available for other filesystems. It is not hard, but it is needed. Maybe it is even better to
move the protection to the filesystem class and just serialize all accesses to a single file system? In this latter case,
direct access to a file should also be protected in the File class.


git-svn-id: https://xp-dev.com/svn/1541UltimateII@251 6fa36889-7ba6-4ba5-89ae-355e570fd296